contact refactor to Long

+++ contact-service/src/main/java/com/ironhack/contactservice/controller/interfaces/IContactController.java
@@ -2,15 +2,14 @@ package com.ironhack.contactservice.controller.interfaces;
 
 import com.ironhack.contactservice.controller.dto.AccountIdDTO;
 import com.ironhack.contactservice.controller.dto.ContactDTO;
-import com.ironhack.contactservice.controller.dto.LeadDTO;
 import com.ironhack.contactservice.model.Contact;
 
 public interface IContactController {
 
     //Get method: get a contact by id
-    ContactDTO getContactDTOById(int id);
+    ContactDTO getContactDTOById(Long id);
     //Post method: create a contact
-    Contact createContact(int id);
+    Contact createContact(Long id);
     //Patch method: update a contact account_id
-    void updateContactAccountId(int id, AccountIdDTO accountIdDTO);
+    void updateContactAccountId(Long id, AccountIdDTO accountIdDTO);
 }
