+++ contact-service/src/main/java/com/ironhack/contactservice/service/impl/ContactService.java
@@ -24,7 +24,7 @@ public class ContactService implements IContactService {
     //===========================================
     //Get methods
     //===========================================
-    public ContactDTO getContactById(int id) {
+    public ContactDTO getContactById(Long id) {
 
         if(contactRepository.findById(id).isEmpty()) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Contact with id " + id + " not found");
@@ -42,7 +42,7 @@ public class ContactService implements IContactService {
     //===========================================
     //Post methods
     //===========================================
-    public Contact createContact(int id) {
+    public Contact createContact(Long id) {
 
         //Get a leadDTO from microservice lead, to create a new contact with the data
         LeadDTO leadDTO = leadClient.getLeadDTOById(id);
@@ -54,7 +54,7 @@ public class ContactService implements IContactService {
     //===========================================
     //Patch methods
     //===========================================
-    public void updateContact(int id, AccountIdDTO accountIdDTO) {
+    public void updateContact(Long id, AccountIdDTO accountIdDTO) {
 
         if(contactRepository.findById(id).isEmpty()) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Contact with id " + id + " not found");
